[Feat] Folder 구조 구현 완료

+++ src/main/java/com/aisip/OnO/backend/service/FolderServiceImpl.java
@@ -6,16 +6,22 @@ import com.aisip.OnO.backend.Dto.Problem.ProblemResponseDto;
 import com.aisip.OnO.backend.entity.Folder;
 import com.aisip.OnO.backend.entity.Problem;
 import com.aisip.OnO.backend.entity.User.User;
+import com.aisip.OnO.backend.exception.FolderNotFoundException;
+import com.aisip.OnO.backend.exception.ProblemNotFoundException;
+import com.aisip.OnO.backend.exception.UserNotFoundException;
 import com.aisip.OnO.backend.repository.FolderRepository;
 import com.aisip.OnO.backend.repository.ProblemRepository;
 import com.aisip.OnO.backend.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @Service
 @RequiredArgsConstructor
 @Transactional
@@ -34,8 +40,10 @@ public class FolderServiceImpl implements FolderService {
         Optional<Folder> optionalRootFolder = folderRepository.findByUserIdAndParentFolderIsNull(userId);
 
         if (optionalRootFolder.isPresent()) {
+            log.info("find root folder id: " + optionalRootFolder.get().getId());
             return findFolder(userId, optionalRootFolder.get().getId());
         } else {
+            log.info("create root folder for userId: " + userId);
             return createRootFolder(userId, "메인");
         }
     }
@@ -44,7 +52,7 @@ public class FolderServiceImpl implements FolderService {
     public FolderResponseDto createRootFolder(Long userId, String folderName) {
         Optional<User> optionalUser = userRepository.findById(userId);
         if (optionalUser.isPresent()) {
-            Folder rootFolder = Folder.builder().name(folderName).user(optionalUser.get()).parentFolder(null).build();
+            Folder rootFolder = Folder.builder().name(folderName).user(optionalUser.get()).build();
 
             folderRepository.save(rootFolder);
 
@@ -58,14 +66,20 @@ public class FolderServiceImpl implements FolderService {
             return findFolder(userId, rootFolder.getId());
         }
 
-        return null;
+        throw new UserNotFoundException("유저를 찾을 수 없습니다!");
     }
 
     @Override
     public FolderResponseDto createFolder(Long userId, String folderName, Long parentFolderId) {
         Optional<User> optionalUser = userRepository.findById(userId);
+
+        log.info("folderName: " + folderName + " parentFolderId : " + parentFolderId);
+
         if (optionalUser.isPresent()) {
-            Folder folder = Folder.builder().name(folderName).user(optionalUser.get()).build();
+            Folder folder = Folder.builder()
+                    .name(folderName)
+                    .user(optionalUser.get())
+                    .build();
 
             if (parentFolderId != null) {
                 Optional<Folder> parentFolder = folderRepository.findById(parentFolderId);
@@ -77,7 +91,7 @@ public class FolderServiceImpl implements FolderService {
             return findFolder(userId, folder.getId());
         }
 
-        return null;
+        throw new UserNotFoundException("유저를 찾을 수 없습니다!");
     }
 
     @Override
@@ -89,18 +103,74 @@ public class FolderServiceImpl implements FolderService {
 
             Folder folder = optionalFolder.get();
 
-            FolderThumbnailResponseDto parentFolder = FolderThumbnailResponseDto.builder().folderId(folder.getParentFolder().getId()).folderName(folder.getParentFolder().getName()).build();
+            FolderThumbnailResponseDto parentFolder = null;
+            if (folder.getParentFolder() != null) {
+                parentFolder = FolderThumbnailResponseDto.builder()
+                        .folderId(folder.getParentFolder().getId())
+                        .folderName(folder.getParentFolder().getName())
+                        .build();
+            }
 
-            List<FolderThumbnailResponseDto> subFolders = folder.getSubFolders().stream().map(subFolder -> {
-                return FolderThumbnailResponseDto.builder().folderId(subFolder.getId()).folderName(subFolder.getName()).build();
-            }).toList();
+            List<FolderThumbnailResponseDto> subFolders = null;
+
+            if (folder.getSubFolders() != null && !folder.getSubFolders().isEmpty()) {
+                subFolders = folder.getSubFolders().stream().map(subFolder -> {
+                    return FolderThumbnailResponseDto.builder()
+                            .folderId(subFolder.getId())
+                            .folderName(subFolder.getName())
+                            .build();
+
+                }).toList();
+            }
 
             List<ProblemResponseDto> problems = problemService.findAllProblemsByFolderId(folderId);
 
-            return FolderResponseDto.builder().folderId(folder.getId()).folderName(folder.getName()).parentFolder(parentFolder).subFolders(subFolders).problems(problems).updateAt(folder.getUpdatedAt()).createdAt(folder.getCreatedAt()).build();
+            return FolderResponseDto.builder()
+                    .folderId(folder.getId())
+                    .folderName(folder.getName())
+                    .parentFolder(parentFolder)
+                    .subFolders(subFolders)
+                    .problems(problems)
+                    .updateAt(folder.getUpdatedAt())
+                    .createdAt(folder.getCreatedAt())
+                    .build();
         }
 
-        return null;
+        throw new FolderNotFoundException("폴더를 찾을 수 없습니다!");
+    }
+
+    @Override
+    public List<FolderThumbnailResponseDto> findAllFolderThumbnailsByUserId(Long userId) {
+
+        List<Folder> folders = folderRepository.findAllByUserId(userId);
+
+        if (folders.isEmpty()) {
+            return new ArrayList<>();
+        }
+
+        return folders.stream().map(folder -> {
+
+            Long parentFolderId = null;
+            if (folder.getParentFolder() != null) {
+                parentFolderId = folder.getParentFolder().getId();
+            }
+
+            List<Long> subFoldersId = new ArrayList<>();
+            List<Folder> subFolders = folder.getSubFolders();
+
+            if (subFolders != null && !subFolders.isEmpty()) {
+                subFoldersId = subFolders.stream()
+                        .map(Folder::getId).toList();
+            }
+
+            return FolderThumbnailResponseDto.builder()
+                    .folderId(folder.getId())
+                    .folderName(folder.getName())
+                    .parentFolderId(parentFolderId)
+                    .subFoldersId(subFoldersId)
+                    .build();
+
+        }).toList();
     }
 
     @Override
@@ -109,14 +179,19 @@ public class FolderServiceImpl implements FolderService {
         if (optionalUser.isPresent()) {
             Optional<Folder> optionalFolder = folderRepository.findById(folderId);
 
+            log.info("folderName: " + folderName + " , parentFolderId : " + parentFolderId);
+
             if (optionalFolder.isPresent()) {
                 Folder folder = optionalFolder.get();
-                folder.setName(folderName);
 
-                Optional<Folder> optionalParentFolder = folderRepository.findById(parentFolderId);
-                Folder parentFolder = optionalParentFolder.orElse(null);
+                if (folderName != null) {
+                    folder.setName(folderName);
+                }
 
-                folder.setParentFolder(parentFolder);
+                if (parentFolderId != null && !(parentFolderId.equals(folderId))) {
+                    Optional<Folder> optionalParentFolder = folderRepository.findById(parentFolderId);
+                    optionalParentFolder.ifPresent(folder::setParentFolder);
+                }
 
                 folderRepository.save(folder);
 
@@ -124,21 +199,81 @@ public class FolderServiceImpl implements FolderService {
             }
         }
 
-        return null;
+        throw new UserNotFoundException("유저를 찾을 수 없습니다!");
     }
 
     @Override
-    public void deleteFolder(Long folderId) {
-        folderRepository.deleteById(folderId);
+    public FolderResponseDto updateProblemPath(Long userId, Long problemId, Long folderId) {
+
+        Optional<User> optionalUser = userRepository.findById(userId);
+        Optional<Problem> optionalProblem = problemRepository.findById(problemId);
+        Optional<Folder> optionalFolder = folderRepository.findById(folderId);
+
+        if (optionalUser.isPresent()) {
+            if (optionalProblem.isPresent() && optionalFolder.isPresent()) {
+                Problem problem = optionalProblem.get();
+                Folder folder = optionalFolder.get();
+
+                problem.setFolder(folder);
+                problemRepository.save(problem);
+
+                return findFolder(userId, folderId);
+            }
+
+            throw new ProblemNotFoundException("문제를 찾을 수 없습니다!");
+        }
+
+        throw new UserNotFoundException("유저를 찾을 수 없습니다!");
     }
 
     @Override
-    public List<FolderThumbnailResponseDto> findAllFolderThumbnailsByUserId(Long userId) {
+    public FolderResponseDto deleteFolder(Long userId, Long folderId) {
+        Optional<Folder> optionalFolder = folderRepository.findById(folderId);
+
+        if (optionalFolder.isPresent() && optionalFolder.get().getUser().getId().equals(userId)) {
+            Folder folder = optionalFolder.get();
+            Long parentFolderId = folder.getParentFolder().getId();
+
+            // 재귀적으로 하위 폴더를 삭제
+            deleteSubFolders(folder);
+            folderRepository.flush();
+
+            if (parentFolderId != null) {
+                log.info("Folder and its subfolders deleted successfully for folderId: " + folderId);
+                folderRepository.deleteById(folderId);
+
+                log.info("root folder id: " + parentFolderId);
+                return findFolder(userId, parentFolderId);
+            } else {
+                throw new FolderNotFoundException("메인 폴더는 삭제할 수 없습니다.");
+            }
+
+        } else {
+            throw new FolderNotFoundException("폴더를 찾을 수 없습니다!");
+        }
+    }
+
+    private void deleteSubFolders(Folder folder) {
+        // 하위 폴더들을 먼저 삭제
+        if (folder.getSubFolders() != null && !folder.getSubFolders().isEmpty()) {
+            for (Folder subFolder : folder.getSubFolders()) {
+                deleteSubFolders(subFolder); // 재귀적으로 하위 폴더 삭제
+            }
+        }
+
+        if (folder.getProblems() != null && !folder.getProblems().isEmpty()) {
+            problemRepository.deleteAll(folder.getProblems()); // 문제들 삭제
+        }
+
+        folder.getSubFolders().clear(); // 하위 폴더 리스트 비우기 (orphanRemoval 보장)
+        folderRepository.delete(folder); // 폴더 삭제
+    }
+
+    @Override
+    public void deleteAllUserFolder(Long userId) {
 
         List<Folder> folders = folderRepository.findAllByUserId(userId);
 
-        return folders.stream().map(folder -> {
-            return FolderThumbnailResponseDto.builder().folderId(folder.getId()).folderName(folder.getName()).build();
-        }).toList();
+        folderRepository.deleteAll(folders);
     }
 }
